# openapi.yaml
openapi: 3.0.3
info:
  title: Dog Trainers Directory API
  description: |
    API for the Dog Trainers Directory platform.
    
    This API provides endpoints for:
    - Public access to trainer listings and search
    - Trainer profile management
    - Admin moderation and management
    - Emergency triage and contacts
    - Authentication with MFA support
    
    ## Authentication
    
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    
    `Authorization: Bearer <access_token>`
    
    ## Rate Limiting
    
    All endpoints are rate limited. Rate limit headers are included in responses:
    
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
    - `Retry-After`: Seconds until retry (only when rate limited)
  version: 1.0.0
  contact:
    name: Dog Trainers Directory Support
    email: support@dogtrainersdirectory.com.au
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.dogtrainersdirectory.com.au/v1
    description: Production server
  - url: https://staging-api.dogtrainersdirectory.com.au/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /suburbs:
    get:
      tags:
        - Public
      summary: List suburbs
      description: Get list of Melbourne suburbs with filtering
      operationId: getSuburbs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/CouncilId'
        - $ref: '#/components/parameters/Region'
        - $ref: '#/components/parameters/Postcode'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuburbsResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /admin/trainers/{id}/status:
    put:
      tags:
        - Admin
      summary: Update trainer status
      description: Update trainer verification and claimed status (admin only)
      operationId: updateTrainerStatus
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrainerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrainerStatusRequest'
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateTrainerStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/reviews/pending:
    get:
      tags:
        - Admin
      summary: Get pending reviews
      description: Get list of pending reviews for moderation (admin only)
      operationId: getPendingReviews
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/CouncilId'
        - $ref: '#/components/parameters/MinRating'
        - $ref: '#/components/parameters/MaxRating'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingReviewsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /admin/featured/queue:
    get:
      tags:
        - Admin
      summary: Get featured queue
      description: Get featured placement queue (admin only)
      operationId: getFeaturedQueue
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/CouncilId'
        - $ref: '#/components/parameters/FeaturedStatus'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeaturedQueueResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /admin/emergency/logs:
    get:
      tags:
        - Admin
      summary: Get emergency logs
      description: Get emergency triage logs (admin only)
      operationId: getEmergencyLogs
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Classification'
        - $ref: '#/components/parameters/MinConfidence'
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmergencyLogsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  parameters:
    Page:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    CouncilId:
      name: council_id
      in: query
      description: Filter by council ID
      required: false
      schema:
        type: string
        format: uuid
...
    Region:
      name: region
      in: query
      description: Filter by region
      required: false
      schema:
        type: string
        enum: [Inner City, Northern, Eastern, South Eastern, Western]
    Postcode:
      name: postcode
      in: query
      description: Filter by postcode
      required: false
      schema:
        type: string
        pattern: '^\\d{4}$'
    MinRating:
      name: min_rating
      in: query
      description: Minimum rating
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 5
    MaxRating:
      name: max_rating
      in: query
      description: Maximum rating
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 5
    TrainerId:
      name: id
      in: path
      description: Trainer ID
      required: true
      schema:
        type: string
        format: uuid
...
    FeaturedStatus:
      name: status
      in: query
      description: Filter by featured status
      required: false
      schema:
        type: string
        enum: [queued, active, expired, refunded, cancelled]
    Classification:
      name: classification
      in: query
      description: Filter by triage classification
      required: false
      schema:
        type: string
        enum: [medical, crisis, stray, normal]
    MinConfidence:
      name: min_confidence
      in: query
      description: Minimum confidence score
      required: false
      schema:
        type: number
        minimum: 0
        maximum: 1
    DateFrom:
      name: date_from
      in: query
      description: Filter by date from
      required: false
      schema:
        type: string
        format: date-time
    DateTo:
      name: date_to
      in: query
      description: Filter by date to
      required: false
      schema:
        type: string
        format: date-time
...
  schemas:
...
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        version:
          type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...
    Trainer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        resource_type:
          type: string
          enum: [trainer, behaviour_consultant, emergency_vet, urgent_care, emergency_shelter]
        locality_id:
          type: string
          format: uuid
        council_id:
          type: string
          format: uuid
        phone:
          type: string
        email:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        age_specialties:
          type: array
          items:
            type: string
        behavior_issues:
          type: array
          items:
            type: string
        service_type_primary:
          type: string
          nullable: true
        service_type_secondary:
          type: array
          items:
            type: string
          nullable: true
        abr_abn:
          type: string
          nullable: true
        emergency_phone:
          type: string
          nullable: true
        emergency_hours:
          type: string
          nullable: true
        emergency_services:
          type: array
          items:
            type: string
          nullable: true
        verified:
          type: boolean
        claimed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Locality:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        council_id:
          type: string
          format: uuid
        region:
          type: string
          enum: [Inner City, Northern, Eastern, South Eastern, Western]
        postcode:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        ux_label:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        moderation_status:
          type: string
          enum: [pending, approved, rejected]
        approved_at:
          type: string
          format: date-time
          nullable: true
        rejected_at:
          type: string
          format: date-time
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FeaturedPlacement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_id:
          type: string
          format: uuid
        council_id:
          type: string
          format: uuid
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [queued, active, expired, refunded, cancelled]
        queue_position:
          type: integer
          nullable: true
        refund_reason:
          type: string
          nullable: true
        refunded_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SuburbsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            suburbs:
              type: array
              items:
                $ref: '#/components/schemas/Locality'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            has_more:
              type: boolean
            council_id:
              type: string
              format: uuid
            region:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    UpdateTrainerStatusRequest:
      type: object
      properties:
        verified:
          type: boolean
        claimed:
          type: boolean
        status_reason:
          type: string
          minLength: 1
          maxLength: 500
    UpdateTrainerStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            trainer:
              $ref: '#/components/schemas/Trainer'
            previous_status:
              type: object
              properties:
                verified:
                  type: boolean
                claimed:
                  type: boolean
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    PendingReviewsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            reviews:
              type: array
              items:
                type: object
                properties:
                  review:
                    $ref: '#/components/schemas/Review'
                  business:
                    $ref: '#/components/schemas/Trainer'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            has_more:
              type: boolean
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    FeaturedQueueResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            placements:
              type: array
              items:
                type: object
                properties:
                  featured_placement:
                    $ref: '#/components/schemas/FeaturedPlacement'
                  business:
                    $ref: '#/components/schemas/Trainer'
                  council:
                    $ref: '#/components/schemas/Council'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            has_more:
              type: boolean
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    EmergencyLogsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            logs:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  user_message:
                    type: string
                  classification:
                    type: string
                    enum: [medical, crisis, stray, normal]
                  confidence_score:
                    type: number
                  ai_response:
                    type: string
                    nullable: true
                  created_at:
                    type: string
                    format: date-time
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            has_more:
              type: boolean
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'