# src/app/api/v1/admin/emergency/logs/route.ts
// ============================================================================
// DTD Phase 2: API Contract - Get Emergency Triage Logs
// File: src/app/api/v1/admin/emergency/logs/route.ts
// Description: Get emergency triage logs
// ============================================================================

import { NextRequest, NextResponse } from 'next/server';
import { supabase, verifyAccessToken, isAdmin } from '@/lib/auth';
import {
  emergencyLogsQuerySchema,
  validateQueryParams,
} from '@/lib/validation';
import {
  handleSupabaseError,
  formatErrorResponse,
  UnauthorizedError,
  ForbiddenError,
} from '@/lib/errors';
import { RateLimiters, getClientIp, checkRateLimitOrThrow, formatRateLimitHeaders } from '@/lib/rate-limit';
import type { EmergencyLogsQuery } from '@/types/api';

/**
 * GET /api/v1/admin/emergency/logs - Get emergency triage logs
 */
export async function GET(request: NextRequest) {
  try {
    // Rate limiting
    const ip = getClientIp(request);
    const rateLimitResult = checkRateLimitOrThrow(RateLimiters.admin, ip);

    // Get authorization header
    const authHeader = request.headers.get('authorization');
    if (!authHeader) {
      throw new UnauthorizedError('Authorization header required');
    }

    const token = authHeader.replace('Bearer ', '');
    const user = await verifyAccessToken(token);

    // Check if user is an admin
    if (!isAdmin(user)) {
      throw new ForbiddenError('Only admins can access this endpoint');
    }

    // Parse and validate query parameters
    const searchParams = request.nextUrl.searchParams;
    const queryParams = Object.fromEntries(searchParams.entries());
    const validatedParams = validateQueryParams(
      emergencyLogsQuerySchema,
      queryParams
    ) as EmergencyLogsQuery;

    // Build Supabase query
    let query = supabase
      .from('triage_logs')
      .select('*');

    // Apply filters
    if (validatedParams.classification) {
      query = query.eq('classification', validatedParams.classification);
    }

    if (validatedParams.min_confidence) {
      query = query.gte('confidence_score', validatedParams.min_confidence);
    }

    if (validatedParams.date_from) {
      query = query.gte('created_at', validatedParams.date_from);
    }

    if (validatedParams.date_to) {
      query = query.lte('created_at', validatedParams.date_to);
    }

    // Apply sorting
    query = query.order('created_at', { ascending: false });

    // Apply pagination
    const page = validatedParams.page || 1;
    const limit = validatedParams.limit || 20;
    const offset = (page - 1) * limit;
    query = query.range(offset, offset + limit - 1);

    // Execute query
    const { data: logs, error, count } = await query;

    if (error) {
      throw handleSupabaseError(error);
    }

    const total = count || 0;
    const hasMore = offset + limit < total;

    // Return response
    return NextResponse.json(
      {
        success: true,
        data: {
          logs,
          total,
          page,
          limit,
          has_more: hasMore,
        },
        meta: {
          timestamp: new Date().toISOString(),
          version: '1.0.0',
        },
      },
      {
        status: 200,
        headers: {
          ...formatRateLimitHeaders(rateLimitResult),
          'Content-Type': 'application/json',
        },
      }
    );
  } catch (error) {
    const apiError = error instanceof Error ? error : new Error('Unknown error');
    return NextResponse.json(
      formatErrorResponse(apiError as any, apiError.message || 'Failed to get emergency logs'),
      {
        status: apiError instanceof UnauthorizedError || apiError instanceof ForbiddenError ? 401 : 500,
        headers: { 'Content-Type': 'application/json' },
      }
    );
  }
}